1.0.2

+++ cj.netos.uc.plugin.core/src/cj/netos/uc/plugin/mapper/TenantAppMapper.java
@@ -62,7 +62,7 @@ public interface TenantAppMapper {
      */
     int updateByPrimaryKey(TenantApp record);
 
-    List<TenantApp> pageApp(@Param(value = "tenantid") String tenantid, @Param(value = "currPage") int currPage, @Param(value = "pageSize") int pageSize);
+    List<TenantApp> pageApp(@Param(value = "tenantid") String tenantid, @Param(value = "currPage") long currPage, @Param(value = "pageSize") int pageSize);
 
     void updateWebsite(@Param(value = "appId") String appId, @Param(value = "website") String website, @Param(value = "loginCbUrl") String loginCbUrl, @Param(value = "logoutCbUrl") String logoutCbUrl);
